+++ src/group/p256.rs
@@ -42,7 +42,7 @@ pub type L = U48;
 
 #[cfg(feature = "p256")]
 impl Group for NistP256 {
-    const SUITE_ID: usize = 0x0003;
+    const SUITE_ID: u16 = 0x0003;
 
     type Elem = ProjectivePoint;
